requirements.txt
addict==2.4.0
dash==2.6.2
huggingface_hub==0.14.1
imageio==2.22.2
lpips==0.1.4
matplotlib==3.7.1
mmcv_full==1.6.2
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
plotly==5.10.0
pycocotools==2.0.5
PyYAML==6.0
scikit_image==0.19.3
scipy==1.10.0
segment_anything==1.0
setuptools==63.4.1
supervision==0.6.0
timm==0.6.13
torch==1.12.1
torch_efficient_distloss==0.1.3
torch_scatter==2.0.9
torchvision==0.13.1
tqdm==4.64.1
transformers==4.29.1
yapf==0.32.0
